requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Templates
jinja2==3.1.3

# HTTP Client for tests
httpx==0.26.0
aiohttp==3.9.3

# PDF Generation
weasyprint==60.2
reportlab==4.0.9

# Security
itsdangerous==2.1.2
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
pydantic[email]==2.6.0
pydantic-settings==2.1.0

# Date handling
python-dateutil==2.8.2
